fix(stream): Move the definition of Needed to stream
